Extracted tabs for diff. generators into new component. Shelfed component for now
